% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_surveys.R
\name{combine_surveys}
\alias{combine_surveys}
\title{Combine serological surveys}
\usage{
combine_surveys(SeroData1, SeroData2)
}
\arguments{
\item{SeroData1}{A first \code{SeroData} object.}

\item{SeroData2}{A second \code{SeroData} object.}
}
\value{
A \code{SeroData} object.
}
\description{
The function \code{combine_surveys} creates a new object of the class \code{SeroData} from two \code{SeroData} objects. In particular, it allows combining serological surveys sampled at different dates.
}
\examples{
 # Combine two simulated surveys, one acquired in 2015, with one-year age categories, 
 # and another one acquired in 1995 with 10-year age categories. 
 Years <- c(1976,1992)
 FOI <- c(0.2,0.3)
 data1 <- simulate_SeroData(sampling_year = 2015, epidemic_years = Years, foi = FOI)
 data2 <- simulate_SeroData(sampling_year = 1995,age_class=10, epidemic_years = Years, foi = FOI)     
 data <- combine_surveys(data1,data2)             
}
\seealso{
\code{\link{SeroData}} Define the format of the serological data.

\code{\link{subset}} Extract a smaller subset of a  \code{SeroData} object.
}
\author{
Nathanael Hoze \email{nathanael.hoze@gmail.com}
}
