% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_information_criteria.R
\name{compute_information_criteria}
\alias{compute_information_criteria}
\title{Compute the AIC, DIC and WAIC}
\usage{
compute_information_criteria(FOIfit, ...)
}
\arguments{
\item{FOIfit}{A \code{FOIfit} object}
}
\value{
A list with the class \code{information_criteria}, which contains the
following items:

\itemize{

\item AIC: The Aikake Information Criterion.

\item DIC: The Deviance Information Criterion. 

\item WAIC: The Wakanabe-Aikake Information Criterion. 

\item MLE: The Maximum-likelihood, estimated as the maximal value of the likelihood in the drawn samples, used in the AIC.

\item k: Number of parameters, used in the AIC.

\item Dbar: Mean deviance, used in the DIC.

\item pD: Effective number of parameters of the model, used in the DIC.

\item pwaic: Estimated effective number of parameters, used in the WAIC.

\item lpd: log pointwise predictive density, used in the WAIC.
}
}
\description{
This function returns the Aikake Information Criterion (AIC), the Deviance Information Criterion (DIC) and the
 Watanabe Aikake Information Criterion (WAIC) from a fit of the class  \code{\link{FOIfit}}.
}
\examples{
data('one_peak_simulation')
model <- FOImodel(type='outbreak', K=1, background=1)
F1  = fit(model = model, data = data )
compute_information_criteria(FOIfit  = F1)

}
\references{
See Gelman et al. Stat Comput (2014) 24:997-1016
DOI 10.1007/s11222-013-9416-2
}
\author{
Nathanael Hoze \email{nathanael.hoze@gmail.com}
}
