% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_posterior.R
\name{correlation_posterior}
\alias{correlation_posterior}
\title{Plot the correlation matrix of the output}
\usage{
correlation_posterior(FOIfit, show_coefficients = TRUE)
}
\arguments{
\item{FOIfit}{An object of the class \code{FOIfit}.}

\item{show_coefficients}{Boolean. If  \code{TRUE}, shows the value of the correlation coefficient on the plotted matrix. Default  = \code{TRUE}.}
}
\value{
The correlation matrix.
}
\description{
Compute and plot the correlation matrix of the parameters obtained from a MCMC fit.  It returns a n-by-n matrix, where n is the number of parameters.
}
\examples{
data('two_peaks_simulation')
model <- FOImodel('outbreak',K=2, background = 1)
F <- fit(data=data, model=model)
M <- correlation_posterior(F)
}
\author{
Nathanael Hoze \email{nathanael.hoze@gmail.com}
}
