% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.FOIfit}
\alias{plot.FOIfit}
\title{Plot the force of infection}
\usage{
\method{plot}{FOIfit}(
  FOIfit,
  mean_only = FALSE,
  individual_samples = 0,
  YLIM = 1,
  XLIM1 = NULL,
  XLIM2 = NULL,
  ...
)
}
\arguments{
\item{FOIfit}{The \code{FOIfit} object to be plotted.}

\item{mean_only}{Boolean. If \code{FALSE}, plot the mean force of infection as well as an envelope of the 95\% credible interval. Default =  \code{FALSE}.}

\item{individual_samples}{Integer. Number of individual samples to be plotted additionally to the mean and the credible interval of the force of infection. The \code{individual_samples} samples are randomly chosen in the chains. Default = 0.}

\item{YLIM}{Upper limit of the y-axis. Default = 1. The lower limit is set to 0.}
}
\value{
A list of \code{ggplot2} objects.
}
\description{
Plot the mean and credible interval of the yearly force of infection, given an object of the class \code{FOIfit}. The function \code{plot} uses the package \code{ggplot2}. It returns a graphical object that can be further modified.
}
\examples{
data <- simulate_SeroData( max_age = 50, epidemic_years = c(1976,1992), foi = c(0.2,0.3))
model <- FOImodel(type='outbreak', K = 2)
Fit <- fit(model = model, data = data)
p <- plot(Fit)
p+ylim(0,1)
p[[1]]$category # the name of the category plotted 


}
\author{
Nathanael Hoze \email{nathanael.hoze@gmail.com}
}
