% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsero-package.R
\docType{package}
\name{Rsero-package}
\alias{Rsero}
\alias{Rsero-package}
\title{Rsero: Estimate the annual force of infection using serological data}
\description{
Fit models of the force of infection to serological data using Bayesian MCMC methods. The package provides a standardized format for handling serological data, contains various serocatalytic models of force of infection, fits the models to the data, and provides tools to analyse and plot the results of the fits.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nathoze/Rsero}
}

}
