% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seropositive.R
\name{seroprevalence.fit}
\alias{seroprevalence.fit}
\title{Plot the fit of the seroprevalence vs. age}
\usage{
seroprevalence.fit(
  FOIfit,
  individual_samples = 0,
  age_class = 10,
  YLIM = 1,
  fill.color = "#d7def3",
  line.color = "#5e6b91",
  mid.age.plot = TRUE,
  ...
)
}
\arguments{
\item{FOIfit}{The \code{FOIfit} object to be plotted.}

\item{individual_samples}{Integer. Number of individual samples to be plotted additionally to the mean and the credible interval of the force of infection. The \code{indivual_samples} samples are randomly chosen in the chains. Default = 0.}

\item{age_class}{Integer. Length of the age groups (in years). Used if \code{show_data = TRUE}. Default = 10.}

\item{YLIM}{Upper limit of the y-axis. Default = 1. The lower limit is set to 0.}

\item{fill.color}{Ribbon color. By default, fill.color  = "#d7def3"}

\item{line.color}{Line color. By default, line.color  = "#5e6b91"}

\item{mid.age.plot}{Boolean. Whether the data is plotted at the mid-point of the age category (TRUE) or at the mean age of the individuals in the data within this category (FALSE). Default = \code{TRUE}.}

\item{show_data}{Boolean. If \code{TRUE}, shows the fraction of seropositive as well. By default, ages are associated in groups of \code{age_class} years.  Default = \code{TRUE}.}
}
\value{
A list of \code{ggplot2} objects.
}
\description{
Plot the mean and credible interval of the posterior of the seroprevalence, given an object of the class \code{FOIfit}. See function \code{\link{seroprevalence}} for plotting the seroprevalence from the data. 
 The function returns graphical objects that can be further modified.
}
\examples{
data <- simulate_SeroData( max_age = 50, epidemic_years = c(1976,1992), foi = c(0.2,0.3))
model <- FOImodel(type='outbreak', background=TRUE, K = 2)

Fit <- fit(model = model, data = data)

# plot the fraction of seropositive, with 2 individual samples
p <- seroprevalence.fit(Fit,  individual_samples = 2,YLIM=0.2)


}
\author{
Nathanael Hoze \email{nathanael.hoze@gmail.com}
}
