% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroprevalence_plot.R
\name{seroprevalence.plot}
\alias{seroprevalence.plot}
\title{Plot the seroprevalence by age}
\usage{
seroprevalence.plot(data, age_cats = 10, YLIM = 1, ...)
}
\arguments{
\item{data}{An object of the class \code{SeroData}.}

\item{age_cats}{Integer. The length in years of the age classes. Default = 10.}

\item{YLIM}{Upper limit of the y-axis. Default = 1. The lower limit is set to 0.}
}
\value{
The mean and 95% confidence interval of the seroprevalence
}
\description{
Plot the seroprevalence by age-class. DESCRIBE CONF INTERVAL AND CATEGORIOES
}
\examples{

dat  = data("one_peak_simulation")
seroprevalence.plot(data = dat)


}
\author{
Nathanael Hoze \email{nathanael.hoze@gmail.com}
}
