% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroprevalence.R
\name{seroprevalence}
\alias{seroprevalence}
\title{Compute the seroprevalence}
\usage{
seroprevalence(data)
}
\arguments{
\item{data}{An object of the class \code{SeroData}.}
}
\value{
The mean and 95% confidence interval of the seroprevalence
}
\description{
Compute the seroprevalence using an object of the class \code{SeroData}. If multiple \code{category} are defined, it will also compute the seroprevalence for each category.
}
\examples{

## A very simple example of a serological survey with three individuals: 
data = SeroData(age_at_sampling = c(10,32,24), Y=c(0,1,1), max_age = 50, age_cats = 1, sampling_year = 2017)
seroprevalence(data = data)


}
\author{
Nathanael Hoze \email{nathanael.hoze@gmail.com}
}
