% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceplot_Rsero.R
\name{traceplot_Rsero}
\alias{traceplot_Rsero}
\title{Markov chain Traceplot}
\usage{
traceplot_Rsero(
  FOIfit,
  pars = NULL,
  include = TRUE,
  unconstrain = FALSE,
  inc_warmup = FALSE,
  window = NULL,
  nrow = NULL,
  ncol = NULL,
  ...
)
}
\arguments{
\item{FOIfit}{A \code{FOIfit} object.}

\item{pars}{A character vector of parameter names. Defaults to relevant parameters of the model.}

\item{include}{Should the parameters given by the pars argument be included (the default) or excluded from the plot? Only relevant if pars is not missing.}

\item{unconstrain}{Should parameters be plotted on the unconstrained space? Defaults to FALSE.}

\item{inc_warmup}{TRUE or FALSE, indicating whether the warmup sample are included in the trace plot; defaults to FALSE.}

\item{window}{A vector of length 2. Iterations between window[1] and window[2] will be shown in the plot. The default is to show all iterations if inc_warmup is TRUE and all iterations from the sampling period only if inc_warmup is FALSE. If inc_warmup is FALSE the iterations specified in window should not include iterations from the warmup period.}

\item{nrow, }{ncol     Passed to facet_wrap.}

\item{...}{Optional arguments to pass to geom_path (e.g. size, linetype, alpha, etc.).}

\item{object}{An instance of class stanfit.}
}
\description{
Draw the traceplot corresponding to one or more Markov chains, providing a visual way to inspect sampling behavior and assess mixing across chains and convergence.
 It is an adaptation of the function \code{\link[rstan]{traceplot}} in the rstan package.
}
\examples{
##'   data <- simulate_SeroData(number_samples = 1000,
  age_class = 1,
  epidemic_years = c(1976,1992),
  foi = c(0.2,0.3))
model <- FOImodel('outbreak', K = 2)
F1 <- fit(model = model, data = data)
traceplot_Rsero(F1)

}
\author{
Nathanael Hoze \email{nathanael.hoze@gmail.com}
}
