% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FOImodel.R
\name{FOImodel}
\alias{FOImodel}
\alias{print.FOImodel}
\title{Model of the force of infection}
\usage{
FOImodel(type = "constant", K = 1, seroreversion = 0, background = 0,
  prioralpha1 = 0, prioralpha2 = 5, priorbeta1 = 0, priorbeta2 = 5,
  priorT1 = 0, priorT2 = 70, priorC1 = 0, priorC2 = 10, priorY1 = 0,
  priorY2 = 100, priorbg1 = 0, priorbg2 = 1, priorRho1 = 0,
  priorRho2 = 10, cat_bg = 0, cat_lambda = 0, ...)

\method{print}{FOImodel}(x, ...)
}
\arguments{
\item{type}{A character with the name of model. The options are: 
\itemize{
  \item 'constant': Constant force of infection
  \item 'outbreak': Series of outbreak modeled with gaussians 
  \item 'independent': Yearly independent values of the force of infection
  \item 'intervention': Piecewise constant force of infection. The number of phases with a constant level is given by the variable \code{K}
  \item 'constantoutbreak': A combination of \code{K} outbreaks with a constant yearly force of infection
}}

\item{K}{integer. An additional parameter used in the outbreak, constantoutbreak and intervention stan models. This parameter is the number of Gaussians used in the model.  Default = \code{NULL}.}

\item{seroreversion}{integer, equal to 0 or 1. If \code{seroreversion=0} the model includes a rate of seroreversion (waning immunity). See the vignette \code{models} for details.  Default = 0.}

\item{background}{integer, equal to 0 or 1. If \code{background=1}  the model includes a background infection probability. See the vignette \code{models} for details.  Default = 0.}

\item{prioralpha1}{First parameter of the uniform prior distribution  of the parameter alpha, used as the intensity of the force of infection in the outbreak and intervention models. Default = 0.}

\item{prioralpha2}{Second parameter of the uniform prior distribution  of the parameter alpha, used as the intensity of the force of infection in the outbreak and intervention models.  Default = 5.}

\item{priorbeta1}{First parameter of the uniform prior distribution  of the parameter beta, used as the spread of the force of infection in the outbreak and intervention models. Default = 0.}

\item{priorbeta2}{Second parameter of the uniform prior distribution  of the parameter beta, used as the spread of the force of infection in the outbreak and intervention models. Default = 1.}

\item{priorT1}{First parameter for the uniform distribution for the T parameter, used as the time of infection in the outbreak and intervention models. T is defined as the number of years between the survey and the outbreak. Default = 0.}

\item{priorT2}{Second parameter for the uniform distribution for the T parameter, used as the time of infection in the outbreak and intervention models. Default = 70.}

\item{priorC1}{First parameter of the uniform prior distribution for the constant force of infection, used in the constant and intervention models. Default = 0.}

\item{priorC2}{Second parameter of the uniform prior distribution for the constant force of infection, used in the constant and intervention models. Default = 10.}

\item{priorY1}{Mean of the normal distribution for the logit of the annual hazard of infection, used in the independent models. Default = 0.}

\item{priorY2}{Standard-deviation of the normal distribution for the logit of the annual hazard of infection, used in the independent models.  Default = 100.}

\item{priorbg1}{First  parameter of the uniform prior distribution for the \code{pB} parameter (background probability of infection) when background = 1.  Default = 0.}

\item{priorbg2}{Second shape parameter of the beta distribution for the  \code{pB} parameter  (background probability of infection) when background = 1. Default = 1.}

\item{priorRho1}{First parameter of the uniform prior distribution for rho, the seroreversion rate used when seroreversion=1. Default = 0.}

\item{priorRho2}{Second parameter of the uniform prior distribution for rho, the seroreversion rate used when seroreversion=1. Default = 10.}

\item{cat_bg}{integer, equal to 0 or 1. If \code{cat_bg=1} a specific value of the background infection probability \code{pB} is attributed to each category (see \code{SeroData}). See the vignette \code{models} for details. Default = 0.}

\item{cat_lambda}{integer, equal to 0 or 1.  If \code{cat_lambda=1} the force of infection varies accross the different categories defined in an objet \code{SeroData}.  See the vignette \code{models} for details. Default = 0.}

\item{...}{Additional arguments (not used).}

\item{x}{A object of the class \code{FOImodel}.}

\item{cat_lambda}{integer, equal to 0 or 1.  If \code{cat_lambda=1} the force of infection varies accross the different categories defined in an objet \code{SeroData}.  See the vignette \code{models} for details. Default = 0.}
}
\value{
A list with the class \code{FOImodel}, which contains the
following items:

\itemize{

\item type: The type of the model.

\item stanname: The name of the stan file used. 

\item estimated_parameters: The number of estimated parameters.

\item priors: a list with the priors.

\item K: the input parameter used in the outbreak and intervention models, if given.

}
}
\description{
This function creates an object of the class \code{FOImodel}. Inputs are the type of the model (required), additional parameters if required by the model, and parameters for the prior distributions. The models to be given as an input are predefined. 
 More details are given in the vignette  \code{models}.
}
\examples{

## A gaussian model, with one gaussian, and background infection
model <- FOImodel(type='outbreak', background = 1, K = 1)

## A gaussian model, with two gaussians, background infection, and user-defined hyperparameters
model <- FOImodel('outbreak', background=1, K = 2, prioralpha1 = 0, prioralpha2 = 1)

## An intervention model, consisting in two constant phases, with seroreversion and user-defined hyperparameters
model <- FOImodel('intervention', K=2, seroreversion=1, priorRho1=0.1)



}
\author{
Nathanael Hoze \email{nathanael.hoze@gmail.com}
}
