% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_credible_intervals.R
\name{parameters_credible_intervals}
\alias{parameters_credible_intervals}
\title{Analyse a FOIfit}
\usage{
parameters_credible_intervals(FOIfit, quants = c(0.025, 0.5, 0.975))
}
\arguments{
\item{FOIfit}{An object of the class \code{FOIfit}.}

\item{quants}{Numeric. Contains the list of the estimated quantiles given between 0 and 1. Default parameters are the 2.5\%, 50 \% and 97.5 \% quantiles given by \code{quants = c(0.025,0.5,0.975)}.}
}
\value{
A dataframe containing the 2.5%, 50% and 97.5% quantiles and the mean of the posterior distributions. If several categories are specified, the force of infection is given for the reference category and the relative force of infection is given for the other categories.
In the case of an outbreak model (outbreak or outbreak+constant), the total force of infection is given by alpha, and the probability of infection by "Outbreak Prob. Inf.". For models with constant phases, the annual probability of infection is given.
}
\description{
Determines the mean and quantiles of the parameters estimated in the fit. By default, it returns the mean, the median and 95 \% credible interval of the parameters.
}
\examples{
  
 data("1peakSimulation")
 model <- FOImodel(type='outbreak', seroreversion=TRUE, K=1)
 options(mc.cores = parallel::detectCores())
 Fit <- fit(data = data, model = model)
 parameters_credible_intervals(Fit)

# an example where other quantiles are specified
 data("1peakSimulation")
 model <- FOImodel(type='outbreak', seroreversion=TRUE, K=1)
 options(mc.cores = parallel::detectCores())
 Fit <- fit(data = data, model = model)
 parameters_credible_intervals(Fit, quants=c(0.25,0.5,0.75))

}
\author{
Nathanael Hoze \email{nathanael.hoze@gmail.com}
}
